+++ tests/testthat/test_wndprb.R
@@ -1,5 +1,11 @@
 context("Wind Speed Probabilities (wndprb)")
 
+# Set timeout options
+opt.timeout <- getOption("rrricanes.http_timeout")
+opt.attempts <- getOption("rrricanes.http_attempts")
+options("rrricanes.http_timeout" = 1)
+options("rrricanes.http_attempts" = 5)
+
 ## ---- 2008, AL ---------------------------------------------------------------
 al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
 
@@ -31,3 +37,7 @@ test_that("Test ep_prblty_stations()", {
 test_that("Test get_wndprb()", {
     expect_identical(al092008.wndprb, df.al092008.wndprb)
 })
+
+# Reset options
+options("rrricanes.http_timeout" = opt.timeout)
+options("rrricanes.http_attempts" = opt.attempts)