+++ NEWS.md
@@ -19,6 +19,7 @@ rrricanes 0.1.2 (2017-06-08)
 
 * Modified numerous regex patterns to ensure data quality.
 * `tidy_fstadv`, `tidy_wr`, `tidy_fcst` and `tidy_fcst_wr` have been added to replaced now-removed `fstadv_split()`.
+* Added loop to make multiple attempts at extracting contents from NHC archives. Options `rrricanes.http_timeout` and `rrricanes.http_attempts` added to give user more control over this. Default is 3 attempts with no more than 5 permitted.
 
 ### BUG FIXES